pyproject.toml
[project]
name = "fathom-rules"
version = "0.3.0"
description = "Deterministic reasoning runtime for AI agents, built on CLIPS via clipspy"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "clipspy>=1.0.6,<1.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/KrakenNet/fathom"
Repository = "https://github.com/KrakenNet/fathom"

[project.scripts]
fathom = "fathom.cli:app"

[project.optional-dependencies]
server = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
]
mcp = [
    "mcp>=1.0",
]
attestation = [
    "pyjwt[crypto]>=2.8",
    "cryptography>=42.0",
]
langchain = [
    "langchain-core>=0.2",
]
crewai = [
    "crewai>=0.80",
]
openai-agents = [
    "openai-agents>=0.1",
]
google-adk = [
    "google-adk>=1.0",
]
cli = [
    "typer[all]>=0.12",
    "rich>=13",
]
metrics = [
    "prometheus_client",
    "prometheus-fastapi-instrumentator",
]
grpc = [
    "grpcio>=1.60",
    "grpcio-tools",
]
fleet = [
    "redis[hiredis]",
]
fleet-pg = [
    "asyncpg",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
    "mkdocs-redirects>=1.2",
    "mike>=2.1",
    "pdoc>=14.0",
    "codespell>=2.2",
    "mkdocs-swagger-ui-tag>=0.6",
]
all = [
    "fathom-rules[server]",
    "fathom-rules[mcp]",
    "fathom-rules[attestation]",
    "fathom-rules[langchain]",
    "fathom-rules[crewai]",
    "fathom-rules[openai-agents]",
    "fathom-rules[google-adk]",
    "fathom-rules[cli]",
    "fathom-rules[metrics]",
    "fathom-rules[grpc]",
    "fathom-rules[fleet]",
    "fathom-rules[fleet-pg]",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "hypothesis>=6.0",
    "ruff",
    "mypy",
    "types-pyyaml>=6.0.12.20260408",
    "httpx>=0.24",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
    "pytest-asyncio>=1.3.0",
    "typer[all]>=0.12",
    "rich>=13",
]

[tool.uv]
package = true

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fathom"]

[project.entry-points."fathom.packs"]
owasp-agentic = "fathom.rule_packs.owasp_agentic"
nist-800-53 = "fathom.rule_packs.nist_800_53"
hipaa = "fathom.rule_packs.hipaa"
cmmc = "fathom.rule_packs.cmmc"

[tool.ruff]
target-version = "py314"
line-length = 99
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "clips.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "grpc.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "redis.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers -ra"
